MVP-0/requirements.txt
# MVP-0 Requirements
# Multi-agent reasoning system dependencies

# Google Generative AI (Gemini)
google-generativeai>=0.5.0

# Google Drive integration (optional, for loading personas)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Python version
python>=3.10

